Rename global vuex module.
