+++ src/components/organisms/MobileNavigation.vue
@@ -3,7 +3,7 @@
     <ul>
       <li v-for="item in menus.main_menu" :key="item.title">
         <router-link
-          :to="item.slug === 'home' ? '/' : item.slug"
+          :to="item.slug === 'home' ? '/' : '/' + item.slug"
           @click.native="toggleNav"
           exact
           >{{ item.title }}</router-link
@@ -19,7 +19,7 @@ import { mapState, mapActions } from 'vuex';
 export default {
   name: 'MobileNavigation',
   computed: {
-    ...mapState('global', {
+    ...mapState('settings', {
       navIsOpen: state => state.navIsOpen,
     }),
     ...mapState('content', {
@@ -27,7 +27,7 @@ export default {
     }),
   },
   methods: {
-    ...mapActions('global', ['toggleNav']),
+    ...mapActions('settings', ['toggleNav']),
   },
 };
 </script>
