+++ src/components/pages/About.vue
@@ -56,7 +56,7 @@ export default {
   },
   mounted: function() {
     this.loading = true;
-    this.$store.dispatch('global/setColorScheme', {
+    this.$store.dispatch('settings/setColorScheme', {
       colorScheme: 'light',
     });
