+++ src/store/modules/settings.js
